pyproject.toml
[project]
name = "dawmind"
version = "0.1.0"
description = "AI agent that controls FL Studio via native API and vision-based GUI automation"
authors = [{ name = "André Elsen" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "google-generativeai>=0.8.0",
    "mss>=9.0.0",
    "pyautogui>=0.9.54",
    "pillow>=11.0.0",
    "pydantic>=2.10.0",
    "toml>=0.10.2",
    "httpx>=0.28.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[project.scripts]
dawmind = "dawmind.cli:main"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["dawmind*"]

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = ["fl_script/"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
